+++ src/helpers/schema/index.js
@@ -1,0 +1,5 @@
+import validationRules from './validationRules';
+
+export default {
+  validationRules,
+};
